+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [v13.1.0] - 2025-03-25
 
+### Added
+
+- **CUMULUS-4048**
+  - Added a session timeout warning modal that pops up five minutes before the session expires
+
 ### Changed
 
 - **CUMULUS-4028**
@@ -21,6 +26,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
+- **CUMULUS-3263**
+  - Updated datestring format to show the correct hours
 - **CUMULUS-4008**
   - Updated the Dockerfile with improved NPM install command to prevent error messages.